Hacky DenseVector rewiring
ClassTags posed a bit of a challenge: You want to keep the original
classtags created by the compiler, as rewiring everything is too
cumbersome, but at the same time you want to bridge the types from
the original class tag to the new types in the dsl. The way to do
this is to have an implicit conversion from ClassTag[T] to ClassTag[U]
as long as there is a LiftEvidence[T, U] in scope.